pyproject.toml
[build-system]
requires = ["setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "mccode-to-kafka"
dependencies = [
    "ess-streaming-data-types>=0.26.1",
    "confluent-kafka>=2.3.0",
    "numpy",
    "importlib_metadata; python_version<'3.8'",
]
dynamic = ["version"]

[project.scripts]
mccode-to-kafka = "mccode_to_kafka.sender:command_line_send"

[project.optional-dependencies]
test = ["pytest", "mock", "flake8-pyproject", "file-writer-control", "python-on-whales", "h5py"]

[tool.setuptools_scm]

[tool.setuptools.packages.find]
where = ["src"]
include = ["mccode_to_kafka"]

[tool.flake8]
max-line-length = 120
